refactor: extract command claim replay
